refactor(credentials): split file IO and quarantine out of the store
Quality pass on the trapdoor fix.

Correctness:
- Quarantine by COPY, not rename. On Windows a file cannot be renamed while
  any handle to it is open, so a peer process mid-read turned the rename
  into EPERM/EACCES/EBUSY — and this path throwing is the crash loop the
  quarantine exists to prevent. Copying leaves the original for the
  caller's ordinary atomic write to replace, which every credential write
  already does, so quarantine adds no new Windows failure mode.
- Reset the live file inside the quarantine step. `deleteSync` returns
  early when the key is absent and `updateSync` when the updater declines;
  neither reaches a write, so each would otherwise re-quarantine the same
  file on the next call, one copy per attempt.
- Give `account.repairSession` the brain-restart IPC timeout. On the 30s
  default the renderer reports "Repair failed" for a repair still running —
  the exact mis-report the restart channel's budget exists to prevent.
- Platform-neutral copy: "this Mac" appeared in three user-facing strings
  on a cross-platform surface, and the doctor row named a keychain for a
  state that is a DPAPI blob on Windows.
- Log the credential repair to the machine sink, not the project logger:
  on a remote-bound project the project logger ships records to the OTHER
  machine, and a credential repair is a fact about this one.

Structure: `credentialStore.ts` had grown to 1888 lines across three
responsibilities. Split along the real seams — `credentialFileIo.ts` (mode
bits, atomic write, the cross-process lock, the stat watcher) and
`credentialStoreQuarantine.ts` (set-aside, marker, retention) — leaving the
store to own key binding and read policy. The lock timeout is now defined
once, where it is enforced, instead of duplicated as a literal.
