+++ apps/ade-cli/src/services/credentials/credentialStore.ts
@@ -3,6 +3,32 @@ import fs from "node:fs";
 import path from "node:path";
 import { resolveMachineAdeLayout } from "../projects/machineLayout";
 import {
+  CREDENTIAL_FILE_LOCK_TIMEOUT_MS,
+  CredentialFileStatWatcher,
+  defaultLockPath,
+  ensureDirMode700,
+  ensureMode600,
+  isEnoent,
+  isEexist,
+  isSamePath,
+  readJsonObject,
+  readJsonObjectAsync,
+  unlinkIfExists,
+  withCredentialFileLock,
+  withOptionalCredentialFileLock,
+  writeFileAtomic,
+} from "./credentialFileIo";
+import {
+  clearQuarantineMarker,
+  deleteQuarantinedStoreFile,
+  quarantineCredentialFile,
+  quarantineHasExpired,
+  readCredentialStoreQuarantine,
+  readQuarantinedStoreFile,
+  writeQuarantineRecord,
+  type CredentialStoreQuarantineRecord,
+} from "./credentialStoreQuarantine";
+import {
   invalidateDefaultOsBoundKeyMaterialCache,
   platformSupportsOsBoundKeyMaterial,
   readDefaultOsBoundKeyMaterial,
@@ -65,23 +91,15 @@ export type CredentialStoreReadFailureReason =
   /** The file exists but is not a recognised credential envelope. */
   | "store_format";
 
-/**
- * A credential file that could not be read and was moved aside so the process
- * that hit it could keep running.
- *
- * `recoverable` means a peer process may still hold the key (an `os`-sealed
- * store quarantined by the brain, which the desktop app can open). Those are
- * merged back automatically by the first store instance that can decrypt them —
- * see `recoverQuarantinedStore`. Anything else is kept purely for diagnostics.
- */
-export type CredentialStoreQuarantineRecord = {
-  version: 1;
-  at: string;
-  /** Basename of the quarantined ciphertext, in the same directory. */
-  file: string;
-  reason: CredentialStoreReadFailureReason;
-  recoverable: boolean;
+// Re-exported so consumers that only care about credential health — `ade
+// doctor`, the desktop account bridge — have one import to reach for.
+export {
+  readCredentialStoreQuarantine,
+  type CredentialStoreQuarantineRecord,
 };
+
+/** Bounds how often a store re-stats the quarantine marker on a clean read. */
+const QUARANTINE_PROBE_INTERVAL_MS = 5_000;
 
 export type SyncCredentialStore = CredentialStore & {
   getSync(key: string): string | null;
@@ -186,36 +204,14 @@ function fileBackedCredentialWriteError(key: string): Error {
   );
 }
 /**
- * How long a writer waits for the credential-file lock before giving up.
- *
- * Exported because cross-process protocols layered on this store have to
- * out-wait it. In particular the account service polls for a peer's rotated
- * refresh token after a definitive `invalid_grant`: if that poll window were
- * shorter than this timeout, a peer that legitimately won the exchange but is
- * still queued behind the lock would have its session declared dead by the
- * loser.
+ * Re-exported from the file-IO layer that actually enforces it, so a caller
+ * pacing itself against the lock cannot drift from the real timeout.
  */
-export const CREDENTIAL_STORE_LOCK_TIMEOUT_MS = 15_000;
-const LOCK_TIMEOUT_MS = CREDENTIAL_STORE_LOCK_TIMEOUT_MS;
-const LOCK_STALE_MS = 10_000;
-const LOCK_RETRY_MS = 25;
+export const CREDENTIAL_STORE_LOCK_TIMEOUT_MS = CREDENTIAL_FILE_LOCK_TIMEOUT_MS;
 const CREDENTIAL_CHANGE_POLL_INTERVAL_MS = 250;
 /** Bounds OS key-material re-reads when a store keeps failing to decrypt. */
 const KEY_MATERIAL_SELF_HEAL_INTERVAL_MS = 30_000;
-/**
- * How long a recoverable quarantine keeps being retried before ADE stops
- * expecting a peer to turn up with the key. Generous on purpose: the peer here
- * is usually "the user opens the desktop app", which can be a fortnight away on
- * a machine that only runs the brain.
- */
-const QUARANTINE_RETENTION_MS = 30 * 24 * 60 * 60 * 1_000;
-/** Bounds how often a store re-stats the quarantine marker on a clean read. */
-const QUARANTINE_PROBE_INTERVAL_MS = 5_000;
 
-type CredentialLockMetadata = {
-  pid?: number;
-  createdAt?: string;
-};
 
 function normalizeKey(key: string): string {
   const normalized = key.trim();
@@ -224,300 +220,6 @@ function normalizeKey(key: string): string {
   return normalized;
 }
 
-function ensureMode600(filePath: string): void {
-  if (process.platform === "win32") return;
-  try {
-    fs.chmodSync(filePath, 0o600);
-  } catch {
-    // Best effort; some filesystems do not support chmod.
-  }
-}
-
-function ensureDirMode700(dirPath: string): void {
-  fs.mkdirSync(dirPath, { recursive: true, mode: 0o700 });
-  if (process.platform === "win32") return;
-  try {
-    fs.chmodSync(dirPath, 0o700);
-  } catch {
-    // Best effort; some filesystems do not support chmod.
-  }
-}
-
-function writeFileAtomic(filePath: string, contents: string | Buffer): void {
-  ensureDirMode700(path.dirname(filePath));
-  const tmpPath = `${filePath}.${process.pid}.${Date.now()}.tmp`;
-  fs.writeFileSync(tmpPath, contents, { mode: 0o600 });
-  ensureMode600(tmpPath);
-  fs.renameSync(tmpPath, filePath);
-  ensureMode600(filePath);
-}
-
-function isEnoent(error: unknown): boolean {
-  return typeof error === "object"
-    && error !== null
-    && "code" in error
-    && (error as { code?: unknown }).code === "ENOENT";
-}
-
-function isEexist(error: unknown): boolean {
-  return typeof error === "object"
-    && error !== null
-    && "code" in error
-    && (error as { code?: unknown }).code === "EEXIST";
-}
-
-/**
- * Windows does not report lock contention as EEXIST the way POSIX does.
- *
- * Deleting a file on Windows only unlinks the name once every open handle to it
- * closes, so between one holder's unlink and the last handle drop the lock name
- * still occupies the directory in a "delete pending" state. A concurrent
- * `open(lockPath, "wx")` against that name fails with a delete-pending or
- * sharing violation, which Node surfaces as EPERM, EACCES or EBUSY instead of
- * EEXIST. Those are the same "someone else holds it, try again" condition, so
- * they have to keep the acquisition loop running; treating them as fatal makes
- * every concurrent credential write a coin flip on Windows.
- */
-function isLockContention(error: unknown): boolean {
-  if (isEexist(error)) return true;
-  if (process.platform !== "win32") return false;
-  if (typeof error !== "object" || error === null || !("code" in error)) return false;
-  const code = (error as { code?: unknown }).code;
-  return code === "EPERM" || code === "EACCES" || code === "EBUSY";
-}
-
-function sleepSync(ms: number): void {
-  Atomics.wait(new Int32Array(new SharedArrayBuffer(4)), 0, 0, ms);
-}
-
-function defaultLockPath(credentialsPath: string): string {
-  return `${credentialsPath}.lock`;
-}
-
-function isSamePath(left: string, right: string): boolean {
-  return path.resolve(left) === path.resolve(right);
-}
-
-type CredentialFileStatSnapshot = {
-  ino: number;
-  mtimeMs: number;
-  size: number;
-} | null;
-
-function readCredentialFileStatSnapshot(filePath: string): CredentialFileStatSnapshot | undefined {
-  try {
-    const stat = fs.statSync(filePath);
-    return { ino: stat.ino, mtimeMs: stat.mtimeMs, size: stat.size };
-  } catch (error: unknown) {
-    if (isEnoent(error)) return null;
-    return undefined;
-  }
-}
-
-function isSameCredentialFileStat(
-  left: CredentialFileStatSnapshot,
-  right: CredentialFileStatSnapshot,
-): boolean {
-  if (left === null || right === null) return left === right;
-  return left.ino === right.ino
-    && left.mtimeMs === right.mtimeMs
-    && left.size === right.size;
-}
-
-class CredentialFileStatWatcher {
-  private previous: CredentialFileStatSnapshot | undefined;
-  private timer: NodeJS.Timeout | null = null;
-
-  constructor(
-    private readonly filePath: string,
-    private readonly listener: () => void,
-    private readonly intervalMs: number | null,
-  ) {}
-
-  start(): void {
-    this.previous = readCredentialFileStatSnapshot(this.filePath);
-    if (this.intervalMs === null) return;
-    this.timer = setInterval(() => this.checkNow(), this.intervalMs);
-    this.timer.unref();
-  }
-
-  checkNow(): void {
-    const current = readCredentialFileStatSnapshot(this.filePath);
-    if (current === undefined) return;
-    if (this.previous === undefined) {
-      this.previous = current;
-      return;
-    }
-    if (isSameCredentialFileStat(current, this.previous)) return;
-    this.previous = current;
-    try {
-      this.listener();
-    } catch {
-      // Credential observers are best-effort; one subscriber must not stop
-      // the watcher or prevent sibling subscribers from seeing the change.
-    }
-  }
-
-  dispose(): void {
-    if (this.timer) clearInterval(this.timer);
-    this.timer = null;
-  }
-}
-
-function parseLockMetadata(raw: string): CredentialLockMetadata {
-  try {
-    const parsed = JSON.parse(raw) as unknown;
-    if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) return {};
-    const record = parsed as Record<string, unknown>;
-    return {
-      pid: Number.isSafeInteger(record.pid) && Number(record.pid) > 0 ? Number(record.pid) : undefined,
-      createdAt: typeof record.createdAt === "string" ? record.createdAt : undefined,
-    };
-  } catch {
-    return {};
-  }
-}
-
-function isProcessRunning(pid: number): boolean {
-  try {
-    process.kill(pid, 0);
-    return true;
-  } catch (error: unknown) {
-    return !(
-      typeof error === "object"
-      && error !== null
-      && "code" in error
-      && (error as { code?: unknown }).code === "ESRCH"
-    );
-  }
-}
-
-function isSameLockStat(left: fs.Stats, right: fs.Stats): boolean {
-  return left.dev === right.dev
-    && left.ino === right.ino
-    && left.size === right.size
-    && left.mtimeMs === right.mtimeMs;
-}
-
-function removeStaleLock(lockPath: string): void {
-  let originalStat: fs.Stats;
-  let originalRaw: string;
-  try {
-    originalStat = fs.statSync(lockPath);
-    if (Date.now() - originalStat.mtimeMs <= LOCK_STALE_MS) return;
-    originalRaw = fs.readFileSync(lockPath, "utf8");
-  } catch {
-    return;
-  }
-
-  const metadata = parseLockMetadata(originalRaw);
-  if (metadata.pid && isProcessRunning(metadata.pid)) return;
-
-  try {
-    const currentStat = fs.statSync(lockPath);
-    if (!isSameLockStat(currentStat, originalStat)) return;
-    if (fs.readFileSync(lockPath, "utf8") !== originalRaw) return;
-    fs.unlinkSync(lockPath);
-  } catch {
-    // Another process won the lock race or removed the stale file first.
-  }
-}
-
-function withCredentialFileLock<T>(lockPath: string, fn: () => T): T {
-  ensureDirMode700(path.dirname(lockPath));
-  const deadline = Date.now() + LOCK_TIMEOUT_MS;
-  let fd: number | null = null;
-
-  while (fd === null) {
-    try {
-      const candidateFd = fs.openSync(lockPath, "wx", 0o600);
-      try {
-        fs.writeFileSync(
-          candidateFd,
-          JSON.stringify({ pid: process.pid, createdAt: new Date().toISOString() }),
-        );
-        fd = candidateFd;
-      } catch (error: unknown) {
-        try {
-          fs.closeSync(candidateFd);
-        } catch {
-          // ignore
-        }
-        try {
-          fs.unlinkSync(lockPath);
-        } catch {
-          // ignore
-        }
-        throw error;
-      }
-      ensureMode600(lockPath);
-    } catch (error: unknown) {
-      if (!isLockContention(error)) throw error;
-      removeStaleLock(lockPath);
-      if (Date.now() >= deadline) {
-        throw new Error("Timed out waiting for ADE credential store lock.", { cause: error });
-      }
-      sleepSync(LOCK_RETRY_MS);
-    }
-  }
-
-  try {
-    return fn();
-  } finally {
-    try {
-      fs.closeSync(fd);
-    } catch {
-      // ignore
-    }
-    try {
-      fs.unlinkSync(lockPath);
-    } catch {
-      // ignore
-    }
-  }
-}
-
-function unlinkIfExists(filePath: string): void {
-  try {
-    fs.unlinkSync(filePath);
-  } catch (error: unknown) {
-    if (!isEnoent(error)) throw error;
-  }
-}
-
-function withOptionalCredentialFileLock<T>(lockPath: string, skippedLockPath: string, fn: () => T): T {
-  if (isSamePath(lockPath, skippedLockPath)) return fn();
-  return withCredentialFileLock(lockPath, fn);
-}
-
-function readJsonObject(filePath: string): Record<string, unknown> | null {
-  try {
-    const raw = fs.readFileSync(filePath, "utf8");
-    const parsed = JSON.parse(raw) as unknown;
-    if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) return null;
-    return parsed as Record<string, unknown>;
-  } catch (error: unknown) {
-    if (isEnoent(error)) return {};
-    throw error;
-  }
-}
-
-async function readJsonObjectAsync(filePath: string): Promise<{
-  value: Record<string, unknown> | null;
-  exists: boolean;
-}> {
-  try {
-    const raw = await fs.promises.readFile(filePath, "utf8");
-    const parsed = JSON.parse(raw) as unknown;
-    if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-      return { value: null, exists: true };
-    }
-    return { value: parsed as Record<string, unknown>, exists: true };
-  } catch (error: unknown) {
-    if (isEnoent(error)) return { value: {}, exists: false };
-    throw error;
-  }
-}
 
 function normalizeStoredCredentialValues(value: unknown): Record<string, string> {
   if (!value || typeof value !== "object" || Array.isArray(value)) return {};
@@ -838,43 +540,6 @@ function retryDecodeWithRefreshedKeyMaterial(args: {
   return retried.ok ? retried : null;
 }
 
-function quarantineMarkerPath(credentialsPath: string): string {
-  return `${credentialsPath}.quarantine.json`;
-}
-
-function isQuarantineRecord(value: unknown): value is CredentialStoreQuarantineRecord {
-  if (!value || typeof value !== "object" || Array.isArray(value)) return false;
-  const record = value as Partial<CredentialStoreQuarantineRecord>;
-  return record.version === 1
-    && typeof record.at === "string"
-    && Number.isFinite(Date.parse(record.at))
-    && typeof record.file === "string"
-    // A marker names a sibling file, never a path: it is read by processes that
-    // may be less trusted than whatever wrote it, and following `../` out of the
-    // secrets directory is not a thing this format needs to support.
-    && record.file.length > 0
-    && !record.file.includes("/")
-    && !record.file.includes("\\")
-    && record.file !== "."
-    && record.file !== ".."
-    && (record.reason === "decrypt_failure"
-      || record.reason === "no_os_key_material"
-      || record.reason === "store_format")
-    && typeof record.recoverable === "boolean";
-}
-
-export function readCredentialStoreQuarantine(
-  credentialsPath: string,
-): CredentialStoreQuarantineRecord | null {
-  try {
-    const parsed = JSON.parse(
-      fs.readFileSync(quarantineMarkerPath(credentialsPath), "utf8"),
-    ) as unknown;
-    return isQuarantineRecord(parsed) ? parsed : null;
-  } catch {
-    return null;
-  }
-}
 
 /**
  * Health of a credential file, read without touching it.
@@ -1290,51 +955,22 @@ export class EncryptedFileCredentialStore implements SyncCredentialStore {
     this.lastReadFailureReason = args.reason;
     if (!args.forWrite) return {};
     // A failed quarantine is the one case that still has to fail closed: if the
-    // ciphertext could not be moved, writing would destroy it.
-    this.quarantineUnreadableStore(args.reason, args.recoverableByPeer, args.error);
+    // ciphertext could not be copied aside, writing over it would destroy it.
+    const record = quarantineCredentialFile({
+      credentialsPath: this.credentialsPath,
+      reason: args.reason,
+      recoverable: args.recoverableByPeer,
+    });
+    if (record) {
+      // Reset the live file here rather than leaving it to the caller's write.
+      // Not every write path reaches one — `deleteSync` returns early when the
+      // key is absent, and `updateSync` when the updater declines — and each of
+      // those would otherwise quarantine the same unreadable file again on the
+      // next call, one copy per attempt.
+      this.writeAll({});
+      this.lastQuarantineProbeAt = 0;
+    }
     return {};
-  }
-
-  /**
-   * Moves the unreadable ciphertext to a timestamped sibling and records a
-   * marker beside it. Caller holds the lock.
-   */
-  private quarantineUnreadableStore(
-    reason: CredentialStoreReadFailureReason,
-    recoverable: boolean,
-    cause: unknown,
-  ): void {
-    const at = new Date();
-    const stamp = at.toISOString().replace(/[:.]/g, "-");
-    const file = `${path.basename(this.credentialsPath)}.quarantined-${stamp}`;
-    const target = path.join(path.dirname(this.credentialsPath), file);
-    try {
-      fs.renameSync(this.credentialsPath, target);
-      ensureMode600(target);
-    } catch (error: unknown) {
-      if (isEnoent(error)) return; // A peer already moved or removed it.
-      throw new Error(
-        "ADE could not set aside the unreadable credential store, so it was left untouched.",
-        { cause: cause ?? error },
-      );
-    }
-    const record: CredentialStoreQuarantineRecord = {
-      version: 1,
-      at: at.toISOString(),
-      file,
-      reason,
-      recoverable,
-    };
-    try {
-      writeFileAtomic(
-        quarantineMarkerPath(this.credentialsPath),
-        `${JSON.stringify(record, null, 2)}\n`,
-      );
-    } catch {
-      // The marker only drives recovery and reporting; losing it must not fail
-      // the write that the quarantine just unblocked.
-    }
-    this.pruneExpiredQuarantineFiles();
   }
 
   /**
@@ -1357,18 +993,19 @@ export class EncryptedFileCredentialStore implements SyncCredentialStore {
   ): Record<string, string> | null {
     const record = this.readQuarantineRecordThrottled();
     if (!record) return null;
-    const markerPath = quarantineMarkerPath(this.credentialsPath);
-    const quarantinedPath = path.join(path.dirname(this.credentialsPath), record.file);
-    const expired = Date.now() - Date.parse(record.at) > QUARANTINE_RETENTION_MS;
-    if (!record.recoverable || expired) {
+    if (!record.recoverable || quarantineHasExpired(record)) {
       // Nothing here will ever be recovered: stop advertising a pending repair,
       // but keep the ciphertext itself for diagnostics.
-      if (expired) this.clearQuarantineMarker(markerPath);
+      if (quarantineHasExpired(record)) this.forgetQuarantine();
       return null;
     }
     let attempt: CredentialDecodeAttempt;
     try {
-      attempt = decodeCredentialStore(readJsonObject(quarantinedPath), machineKey, material);
+      attempt = decodeCredentialStore(
+        readQuarantinedStoreFile(this.credentialsPath, record),
+        machineKey,
+        material,
+      );
     } catch {
       return null;
     }
@@ -1376,7 +1013,9 @@ export class EncryptedFileCredentialStore implements SyncCredentialStore {
       // This process HAS the OS material the quarantine was waiting for and
       // still cannot open the file — so no peer will. Say so, or every surface
       // keeps telling the user to open an app that has already tried.
-      if (material) this.demoteQuarantineToUnrecoverable(record, markerPath);
+      if (material) {
+        writeQuarantineRecord(this.credentialsPath, { ...record, recoverable: false });
+      }
       return null;
     }
     const merged = { ...values };
@@ -1391,26 +1030,12 @@ export class EncryptedFileCredentialStore implements SyncCredentialStore {
       // The quarantined copy is live credential ciphertext that this process can
       // decrypt. Once its contents are back in the store, keeping it is only
       // extra secret material at rest.
-      unlinkIfExists(quarantinedPath);
-      this.clearQuarantineMarker(markerPath);
+      deleteQuarantinedStoreFile(this.credentialsPath, record);
+      this.forgetQuarantine();
     } catch {
       return changed ? merged : null;
     }
     return changed ? merged : null;
-  }
-
-  private demoteQuarantineToUnrecoverable(
-    record: CredentialStoreQuarantineRecord,
-    markerPath: string,
-  ): void {
-    try {
-      writeFileAtomic(
-        markerPath,
-        `${JSON.stringify({ ...record, recoverable: false }, null, 2)}\n`,
-      );
-    } catch {
-      // Reporting-only state.
-    }
   }
 
   private readQuarantineRecordThrottled(): CredentialStoreQuarantineRecord | null {
@@ -1425,41 +1050,9 @@ export class EncryptedFileCredentialStore implements SyncCredentialStore {
     return readCredentialStoreQuarantine(this.credentialsPath);
   }
 
-  private clearQuarantineMarker(markerPath: string): void {
-    try {
-      unlinkIfExists(markerPath);
-    } catch {
-      // Reporting-only state; a stale marker is not worth failing a read for.
-    }
+  private forgetQuarantine(): void {
+    clearQuarantineMarker(this.credentialsPath);
     this.lastQuarantineProbeAt = 0;
-  }
-
-  /**
-   * Drops quarantined ciphertext nobody came back for. Diagnostics are worth a
-   * month, not forever — these files are undecryptable-by-anyone in the corrupt
-   * case and decryptable-by-a-peer in the recoverable one, and neither should
-   * accumulate one copy per failed startup.
-   */
-  private pruneExpiredQuarantineFiles(): void {
-    const dir = path.dirname(this.credentialsPath);
-    const prefix = `${path.basename(this.credentialsPath)}.quarantined-`;
-    let entries: string[];
-    try {
-      entries = fs.readdirSync(dir);
-    } catch {
-      return;
-    }
-    const active = readCredentialStoreQuarantine(this.credentialsPath)?.file ?? null;
-    for (const entry of entries) {
-      if (!entry.startsWith(prefix) || entry === active) continue;
-      try {
-        const stat = fs.statSync(path.join(dir, entry));
-        if (Date.now() - stat.mtimeMs <= QUARANTINE_RETENTION_MS) continue;
-        fs.unlinkSync(path.join(dir, entry));
-      } catch {
-        // Best effort housekeeping.
-      }
-    }
   }
 
   /**
